desktop-app/requirements.txt
openai-whisper
torch
tkinterdnd2
ttkbootstrap
